fix: break circular import by moving password auth to socratic_system
PROBLEM:
Circular dependency between API and domain layers:
  socrates_api/database imports from socratic_system
  socratic_system/ui imports from socrates_api/auth
  This creates: socratic_system -> socrates_api -> socratic_system (circular)

ROOT CAUSE:
Password hashing utilities (hash_password, verify_password) were in API layer
but also needed by domain layer CLI commands, creating the circular dependency.

SOLUTION:
Moved password module to foundational layer where it belongs:
  - Created socratic_system/auth/password.py with PasswordManager
  - Updated socratic_system/ui/commands/user_commands.py to import from there
  - Made socrates_api/auth/password.py a re-export for backwards compatibility

RESULT:
Clean unidirectional dependency graph:
  socratic_system (foundation)
    └── auth/password.py (used by UI and API)

  socrates_api (API layer, depends on socratic_system)
    └── auth/password.py (re-exports from socratic_system)

VERIFICATION:
✓ All 4 modified files compile successfully
✓ Circular import eliminated
✓ Backwards compatibility maintained (existing imports still work)
✓ Password functionality verified
✓ Both CLI and API use same password implementation

This is correct architecture because:
- Password utilities are foundational infrastructure (like database)
- Both CLI and API depend on them, not the other way around
- No module should depend on the API layer (API depends on domain, not reverse)
