+++ socratic_system/ui/commands/user_commands.py
@@ -7,9 +7,9 @@ from typing import Any, Dict, List
 
 from colorama import Fore, Style
 
-# Import password verification from API module
+# Import password verification from foundational auth module
 # This ensures CLI and API use identical password hashing (bcrypt)
-from socrates_api.auth.password import hash_password, verify_password
+from socratic_system.auth import hash_password, verify_password
 
 from socratic_system.models import User
 from socratic_system.ui.commands.base import BaseCommand
